refactor: simplify header cell structure
