+++ src/react/table-app/header-cell-edit/base-menu.tsx
@@ -3,7 +3,7 @@ import MenuItem from "src/react/shared/menu-item";
 import Padding from "src/react/shared/padding";
 import Stack from "src/react/shared/stack";
 import { CellType, SortDir } from "src/shared/types/types";
-import { SubmenuType } from "../../types";
+import { SubmenuType } from "./types";
 import { useCompare, useInputSelection } from "src/shared/hooks";
 import { getDisplayNameForCellType } from "src/shared/table-state/display-name";
 import { css } from "@emotion/react";
