--- src/react/table-app/header-cell/components/HeaderMenu/styles.css
@@ -1,3 +1,0 @@
-.NLT__header-menu {
-	color: var(--text-normal);
-}
